// src/app/pages/model/inv/inv_brands.model.ts
// tslint:disable-next-line:class-name
export interface inv_brands {
  brand_id: number;
  brand_name: string;
  brand_type: number;
  brand_type_name: string;
  created_date: any;
  created_by: number;
  modified_date: any;
  modified_by: number;
  session_id: number;
  session_name: string;
  status: string;
  deleted: boolean | false ;
}